chore: address pr comments
