+++ src/lib/helpers.test.ts
@@ -23,7 +23,6 @@ import {
   validateHash,
   validateCapabilityNames,
   ValidationError,
-  determineRbacMode,
 } from "./helpers";
 import { sanitizeResourceName } from "../sdk/sdk";
 import * as fc from "fast-check";
@@ -1483,35 +1482,5 @@ describe("matchesRegex", () => {
     const testString = "invalid-email.com";
     const result = matchesRegex(new RegExp(pattern).source, testString);
     expect(result).toBe(false);
-  });
-});
-
-describe("determineRbacMode", () => {
-  test("should allow CLI options to overwrite module config", () => {
-    const opts = { rbacMode: "admin" };
-    const cfg = { pepr: { rbacMode: "scoped" } };
-    const result = determineRbacMode(opts, cfg);
-    expect(result).toBe("admin");
-  });
-
-  test('should return "admin" when cfg.pepr.rbacMode is provided and not "scoped"', () => {
-    const opts = {};
-    const cfg = { pepr: { rbacMode: "admin" } };
-    const result = determineRbacMode(opts, cfg);
-    expect(result).toBe("admin");
-  });
-
-  test('should return "scoped" when cfg.pepr.rbacMode is "scoped"', () => {
-    const opts = {};
-    const cfg = { pepr: { rbacMode: "scoped" } };
-    const result = determineRbacMode(opts, cfg);
-    expect(result).toBe("scoped");
-  });
-
-  test("should default to admin when neither option is provided", () => {
-    const opts = {};
-    const cfg = { pepr: {} };
-    const result = determineRbacMode(opts, cfg);
-    expect(result).toBe("admin");
   });
 });
