+++ src/lib/helpers.ts
@@ -394,19 +394,3 @@ export function replaceString(str: string, stringA: string, stringB: string) {
   const regExp = new RegExp(escapedStringA, "g");
   return str.replace(regExp, stringB);
 }
-
-// determineRbacMode determines the RBAC mode to use based on the cli and the module's config
-export function determineRbacMode(opts: { rbacMode?: string }, cfg: { pepr: { rbacMode?: string } }): string {
-  // CLI overrides the module's config
-  if (opts.rbacMode) {
-    return opts.rbacMode;
-  }
-
-  // if rbacMode is defined and not scoped, return admin
-  if (cfg.pepr.rbacMode && cfg.pepr.rbacMode !== "scoped") {
-    return "admin";
-  }
-
-  // if nothing is defined return admin, else return scoped
-  return cfg.pepr.rbacMode || "admin";
-}
